Time: 1 ms (86.96%), Space: 41.1 MB (42.59%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.java
@@ -2,20 +2,20 @@ class Solution {
     public int mySqrt(int n) {
         if (n == 0 || n == 1) return n;
 
-        int low = 1;
-        int high = n;
+        long low = 1;
+        long high = n;
         int ans = 0;
 
         while (low <= high) {
-            int mid = (low + high) / 2;
-            long square = (long) mid * mid;
+            long mid = (low + high) / 2;
+            long square = mid * mid;
 
             if (square == n) {
-                return (int)mid;
+                return (int) mid;
             } else if (square > n) {
                 high = mid - 1;
             } else {
-                ans = mid;
+                ans = (int) mid;
                 low = mid + 1;
             }
         }
